working on the main.py and the fishing_game.py code
The changes made in the past hour include:

Updated the file paths for the intro and game over images to be in the "images" folder.
Modified the show_intro_screen() method to load the "start.jpg" image from the "images" folder.
Modified the show_game_over_screen() method to load the "end.jpg" image from the "images" folder.
Make sure you have the "start.jpg" and "end.jpg" images located in the "images" folder relative to the location of the fishing_game.py file.