make the game play better - move the man to the top of the screen
I have made the following changes:

Fixed the error where spawn_fish method was missing in the Game class.
Passed the font argument to the FishAnimation constructor.
Updated the import statement in main.py to import fishing_game instead of game.
To summarize the changes made in the last hour:

Added 10 levels of increasing difficulty using the sigmoid function.
Implemented a menu system with options to play the game or exit.
Added a loading screen before the intro screen.
Improved the fish movement using more realistic behavior.
Updated the fishing man's position to always be at the top of the screen.
Implemented a casting animation for the fishing man using the space bar.
Enhanced the game visuals by adding ASCII art for the fishing man and images for the intro and game over screens.
Fixed various bugs and errors